refactor(telegram): rename addSubscriber back to handleUpdate
